refactor: move Rust tests to separate test files
Move all unit tests from inline #[cfg(test)] modules to separate files
in the rust/tests/ directory, keeping source files clean and focused.

Tests moved:
- link.rs tests -> tests/link_tests.rs
- lino_link.rs tests -> tests/lino_link_tests.rs
- parser.rs tests -> tests/parser_tests.rs
- changes_simplifier.rs tests -> tests/changes_simplifier_tests.rs
- link_storage.rs tests -> tests/link_storage_tests.rs
- query_processor.rs tests -> tests/query_processor_tests.rs

All 36 tests pass.
